services/api/requirements.txt
fastapi>=0.104.0
python-multipart>=0.0.9
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
minio>=7.2.0
anyio>=4.0.0
redis[hiredis]>=5.0.0
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1
cryptography>=42.0.0
aiosmtplib>=3.0.0
rrule>=3.0,<4.0
alembic>=1.13.0
webpush>=1.0.0
